Further separated out strings, and moved all data types to load before any API
