+++ include/con4m/datatypes.h
@@ -1,0 +1,14 @@
+#pragma once
+
+#include <con4m/datatypes/kargs.h>
+#include <con4m/datatypes/objects.h>
+#include <con4m/datatypes/colors.h>
+#include <con4m/datatypes/codepoints.h>
+#include <con4m/datatypes/styles.h>
+#include <con4m/datatypes/strings.h>
+#include <con4m/datatypes/lists.h>
+#include <con4m/datatypes/types.h>
+#include <con4m/datatypes/grids.h>
+#include <con4m/datatypes/buffers.h>
+#include <con4m/datatypes/io.h>
+#include <con4m/datatypes/crypto.h>
