+++ include/con4m/datatypes/codepoints.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <con4m.h>
+
+typedef int32_t codepoint_t;
