+++ include/con4m/datatypes/colors.h
@@ -1,0 +1,10 @@
+#pragma once
+
+#include <con4m.h>
+// Temporary.
+typedef int32_t color_t;
+
+typedef struct {
+    const char *name;
+    int32_t rgb;
+} color_info_t;
