--- include/con4m/dict.h
@@ -1,2 +1,0 @@
-#include <con4m.h>
-#pragma once
