+++ include/con4m/strbase.h
@@ -1,0 +1,1 @@
+#pragma once
