+++ src/con4m/ansi.c
@@ -298,12 +298,10 @@ utf8_ansi_render(const utf8_t *s, FILE *outstream)
 	    break;
 	}
     }
+
     if (style_state == U8_STATE_IN_STYLE ||
 	style_state == U8_STATE_DEFAULT_STYLE) {
-	if (current_style != 0) {
-	    ansi_render_style_end(outstream);
-	}
-    }
+     }
 }
 
 // This will have to convert characters to utf-8, since terminals
