+++ src/con4m/object.c
@@ -193,6 +193,13 @@ const con4m_dt_info builtin_type_info[CON4M_NUM_BUILTIN_DTS] = {
 	 .alloc_len = sizeof(render_style_t),
 	 .ptr_info  = (uint64_t *)&rs_pmap,
 	 .vtable    = &render_style_vtable
+     },
+     {
+	 .name      = "hash",
+	 .typeid    = T_SHA,
+	 .alloc_len = sizeof(sha_ctx),
+	 .ptr_info  = GC_SCAN_ALL,
+	 .vtable    = &sha_vtable
      }
 };
