+++ src/con4m/string.c
